+++ doc/CHANGES.rst
@@ -51,6 +51,8 @@ Features Added
 Bugs Fixed
 ++++++++++
 
+- LP #414757: Don't send a request closed event from a cloned request.
+
 - LP #418454: FTP server did not work with Python 2.6.X
 
 - Fixed issue with sending text containing ':' from MailHost.